[bug fix] Important Bug Fix: 1.when the transfered key TTL more than 2147483647/1000, (max int in ms),the key's TTL will be wrong or the key will failed to restore. 2.when a key is transfered failed, the old transfer tool will continue ,which may cause data lost.
Now when transfering ,if there is some key transfer failed, all the transfer will stop, and fix the bug above.

+++ transfer/hiredis.h
@@ -234,7 +234,7 @@ void *redisCommand(redisContext *c, const char *format, ...);
 void *redisCommandArgv(redisContext *c, int argc, const char **argv, const size_t *argvlen);
 
 /* support for restore command */
-void *redisRestoreCommand (redisContext *c, char *key, size_t keylen, int ttl, char * buf, size_t buflen);
+void *redisRestoreCommand (redisContext *c, char *key, size_t keylen, long long ttl, char * buf, size_t buflen);
 
 #ifdef __cplusplus
 }
